Fix: 코드 리뷰 사항 반영 - 데이터 존재 유무 확인 방식을 .exists()로 통일 - rating에서 참조하는 테이블에 직접 접근하도록 수정 - models에서 created_at, updated_at 수정 - 별점 수정 시 기존과 같은 별점이 들어올 경우 수정이 되지 않도록 수정

+++ archive/views.py
@@ -14,12 +14,12 @@ class RatingView(View):
     @id_auth
     def post(self, request, content_pk):
         try:
-            data   = json.loads(request.body)
-            user = request.user
+            data    = json.loads(request.body)
+            user    = request.user
             content = Content.objects.get(id=content_pk)
             rating  = data['rating']
 
-            if Rating.objects.filter(user = user, content = content):
+            if Rating.objects.filter(user = user, content = content).exists():
                 return JsonResponse({"message": "ALREADY_EIXST"}, status = 400)
 
             Rating.objects.create(user = user, content = content, rating = rating)
@@ -31,37 +31,38 @@ class RatingView(View):
             return JsonResponse({"message": "KEY_ERROR"}, status = 400)
 
     def get(self, request, content_pk):
-        if not Content.objects.filter(id = content_pk):
-            return JsonResponse({"message": "INVALID_CONTENT_ID"}, status = 400)
-        ratings = Rating.objects.filter(content = content_pk)
-        results = []
-        rating_list = []
+        if Content.objects.filter(id = content_pk).exists():
+            ratings     = Rating.objects.filter(content = content_pk)
+            results     = []
+            rating_list = []
         
-        for rating in ratings:
-            results.append(
-                {
-                    "id"      : rating.id,
-                    "user"    : User.objects.get(id=rating.user_id).username,
-                    "content" : Content.objects.get(id=rating.content_id).title_korean,
-                    "rating"  : rating.rating
-                }
-            )
-            rating_list.append(rating.rating)        
-        average_rating = sum(rating_list)/len(rating_list)
-
-        return JsonResponse({"result": results, "average_rating": average_rating}, status = 200)
+            for rating in ratings:
+                results.append(
+                    {
+                        "id"      : rating.id,
+                        "user"    : rating.user.username,
+                        "content" : rating.content.title_korean,
+                        "rating"  : rating.rating
+                    }
+                )
+                rating_list.append(rating.rating)        
+            average_rating = sum(rating_list)/len(rating_list)
+            return JsonResponse({"result": results, "average_rating": average_rating}, status = 200)
+        return JsonResponse({"message": "INVALID_CONTENT_ID"}, status = 400)
 
     @id_auth
     def patch(self, request, content_pk):
         try:
-            data = json.loads(request.body)
-            user = request.user
+            data    = json.loads(request.body)
+            user    = request.user
             content = Content.objects.get(id=content_pk)
-            rating = data['rating']
+            rating  = data['rating']
 
             patch_object = Rating.objects.get(user = user, content = content)
+
+            if patch_object.rating == float(rating):
+                return JsonResponse({"message": "SAME_RATING_SCORE"}, status = 400)
             patch_object.rating = rating
-            patch_object.updated_at = timezone.now()
             patch_object.save()
             return JsonResponse({"message": "RATING_UPDATED"}, status = 201)
 
@@ -72,17 +73,14 @@ class RatingView(View):
         except Rating.DoesNotExist:
             return JsonResponse({"message": "INVALID_RATING"}, status = 400)
 
-
     @id_auth
     def delete(self, request, content_pk):
-        try:
-            user = request.user
-            content = Content.objects.get(id=content_pk)
+        user = request.user
+        content = Content.objects.get(id=content_pk)
 
-            assert Rating.objects.filter(user = user, content = content), "NOT_RATED"
-            
+        if Rating.objects.filter(user = user, content = content).exists():
             Rating.objects.get(user = user, content = content).delete()
             return JsonResponse({"message": "RATING_DELETED"}, status = 203)
+        return JsonResponse({"message": "NOR_RATED"}, status = 400)
+            
 
-        except AssertionError as e:
-            return JsonResponse({"message": f"{e}"}, status = 400)
